backend/requirements.txt
fastapi
numpy
tensorflow
Pillow
prometheus_client
evidently
pandas
python-multipart
psycopg2
pydantic
mlflow